The shiny resetter
To start: once you run the code it will wait for the input of 'g' on the keyboard so you'll need to go to be on the start screen of the game and then press g. It will continue until you press 'h' or until it finds a shiny starter. You must save in front of where you get the pokemon and save your game *It will not do the beginning for you* . This uses only image recognition and clicks of button so it doesn't have dedicated gui(s) nor other functions other than the stated purpose.